requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
PyJWT==2.4.0
mysqlclient==2.1.0
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
uWSGI==2.0.20
boto3==1.24.16
drf-yasg==1.20.0
